// cli-wallet/src/types.ts
export interface KeypairData {
  publicKey: Uint8Array;
  secretKey: Uint8Array;
}